Moved IEventEnvelopeTransport and implementations under, IpcReceiptTransport namespace to not confuse bi-directional (with receipts) and one-way IPC calls
