+++ src/Lite.EventIpc.Tests/IpcReceiptedTransporters/MemoryMappedTests.cs
@@ -5,8 +5,8 @@
 using System;
 using System.Runtime.Versioning;
 using System.Threading.Tasks;
+using Lite.EventAggregator.IpcReceiptTransport;
 using Lite.EventAggregator.Tests.Models;
-using Lite.EventAggregator.IpcTransport;
 
 namespace Lite.EventAggregator.Tests.IpcReceiptedTransporters;
 
@@ -65,4 +65,5 @@ public class MemoryMappedTests : BaseTestClass
     Assert.AreEqual(PayloadRequest + PayloadResponse, resp.Message);
   }
 }
+
 #endif
