+++ src/Lite.EventIpc/IEventAggregator.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Lite.EventAggregator.IpcReceiptTransport;
 using Lite.EventAggregator.IpcTransport;
 
 namespace Lite.EventAggregator;
