+++ src/SampleApp/IpcTransporters/NamedPipesDemo.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Threading.Tasks;
 using Lite.EventAggregator;
-using Lite.EventAggregator.IpcTransport;
+using Lite.EventAggregator.IpcReceiptTransport;
 
 namespace SampleApp.IpcTransporters;
 
@@ -35,4 +35,5 @@ public class NamedPipesDemo
 
   public record Pong(string Message);
 }
+
 #endif
