+++ src/SampleApp/IpcTransporters/TcpDemo.cs
@@ -6,7 +6,7 @@ using System;
 using System.Net;
 using System.Threading.Tasks;
 using Lite.EventAggregator;
-using Lite.EventAggregator.IpcTransport;
+using Lite.EventAggregator.IpcReceiptTransport;
 
 namespace SampleApp.IpcTransporters;
 
@@ -44,4 +44,5 @@ public class TcpDemo
 
   public record Pong(string Message);
 }
+
 #endif
